cleanup: split SettingsView.swift into focused files
SettingsView.swift (5207 lines) split into:
- ViewHelpers.swift — CopyButton, svgToWhite, Color hex helpers, blockStatusColor, simpleRelativeTime, Markdown renderer
- ScriptDetailView.swift — ScriptDetailView and CardColorSchemePreview (~1700 lines)
- ScriptCatalogView.swift — ScriptCatalogView, LocalScriptRow, CatalogEntryRow (~340 lines)
- MacFeedView.swift — FeedFilter, MacFeedView, ExitCodeCell, FeedDetailCell, FeedEventRow, FeedEventDetailSheet (~493 lines)

SettingsView.swift reduced from 5207 to ~2473 lines. Build verified passing.
